+++ classes/text_filter.php
@@ -30,9 +30,12 @@
  * The way the filter works is as follows:
  *
  *    - look for multilang blocks in the text.
- *    - if there exists texts in the currently active language, print them.
- *    - else, if there exists texts in the current parent language, print them.
- *    - else, if there exists texts in the language 'other', print them.
+ *    - if there exist texts in the currently active language, print them.
+ *    - else, if there exist texts in the parent language(s) of the
+ *      currently active language, unless the parent language is 'en',
+ *      print them (default "parent language behaviour", this is now
+ *      configurable)
+ *    - else, if there exist texts in the language 'other', print them.
  *    - else, don't print any text inside the lang block (this is a change
  *      from previous filter versions behaviour!!!!)
  *
@@ -78,6 +81,11 @@ class text_filter extends \filter_multilang2_base_text_filter {
     protected static $parentcache;
 
     /**
+     * @var object An instance of the string_manager class.
+     */
+    protected static $stringmanager;
+
+    /**
      * @var string The langauge we are currently using to filter multilang blocks.
      *             It can be either the user current language, or the language 'other'
      */
@@ -90,6 +98,19 @@ class text_filter extends \filter_multilang2_base_text_filter {
      *           exists).
      */
     protected $replacementdone;
+
+    /**
+     * @var string What parent languages behaviour is configured for the plugin.
+     */
+    protected $parentlangbehaviour;
+
+    /**
+     * Constructor - Make sure we use the configured parent languages behaviour.
+     */
+    public function __construct() {
+        $this->parentlangbehaviour = get_config('filter_multilang2', 'parentlangbehaviour');
+        self::$stringmanager = get_string_manager();
+    }
 
     /**
      * Filter text before changing format to HTML.
@@ -142,6 +163,21 @@ class text_filter extends \filter_multilang2_base_text_filter {
     }
 
     /**
+     * Reset the parent languages cache.
+     *
+     * We need to reset the parent languages cache every time we
+     * switch from one parent languages behaviour to another. Because
+     * different behaviours may build different parent languages
+     * hierarchies.
+     *
+     * Specially useful (but not only) for unit tests.
+     *
+     */
+    public static function reset_parentcache () :void {
+        self::$parentcache = ['other' => []];
+    }
+
+    /**
      * This function filters the received text based on the language
      * tags embedded in the text, and the current user language or
      * 'other', if present.
@@ -163,13 +199,24 @@ class text_filter extends \filter_multilang2_base_text_filter {
         }
 
         if (!isset(self::$parentcache)) {
-            self::$parentcache['other'] = [];
+            self::reset_parentcache();
         }
 
         $this->replacementdone = false;
         $currlang = current_language();
-        if (!array_key_exists($currlang, self::$parentcache)) {
-            $parentlangs = get_string_manager()->get_language_dependencies($currlang);
+        if ($this->parentlangbehaviour === 'never') {
+            self::reset_parentcache();
+        } else if (!array_key_exists($currlang, self::$parentcache)) {
+            $parentlangs = self::$stringmanager->get_language_dependencies($currlang);
+            if ($this->parentlangbehaviour === 'include_en') {
+                if (count($parentlangs) > 0) {
+                    $realrootisen = self::$stringmanager->get_string('parentlanguage', 'langconfig',
+                                                                     null, $parentlangs[0]);
+                    if ($realrootisen === 'en') {
+                        $parentlangs = array_merge(array('en'), $parentlangs);
+                    }
+                }
+            }
             self::$parentcache[$currlang] = $parentlangs;
         }
 
@@ -230,7 +277,11 @@ class text_filter extends \filter_multilang2_base_text_filter {
          */
         $blocklangs = explode(',', str_replace(' ', '', str_replace('-', '_', strtolower($langblock[1]))));
         $blocktext = $langblock[2];
-        $parentlangs = self::$parentcache[$replacelang];
+        if ($this->parentlangbehaviour === 'never') {
+            $parentlangs = array();
+        } else {
+            $parentlangs = self::$parentcache[$replacelang];
+        }
         foreach ($blocklangs as $blocklang) {
             /* We don't check for empty values of $blocklang as they simply don't
              * match any language and they don't produce any errors or warnings.